+++ experimental/packages/sdk-logs/src/Logger.ts
@@ -39,9 +39,7 @@ export class Logger implements logsAPI.Logger {
   }
 
   public emit(logRecord: logsAPI.LogRecord): void {
-    const currentContext = this._loggerConfig.includeTraceContext
-      ? context.active()
-      : undefined;
+    const currentContext = logRecord.context || context.active();
     /**
      * If a Logger was obtained with include_trace_context=true,
      * the LogRecords it emits MUST automatically include the Trace Context from the active Context,